Finally! I've got two boxes for the scores
They needed to have minimal colors and stuff from their parent
so there is no background colour for the parent, but there needed to be
margins and justification of the content.
Then it was just tinkering with the colours to get them looking ok.
Still need to figure out the correct font type though.
Now on to the buttons!